+++ mitra_api/src/server.rs
@@ -131,7 +131,6 @@ pub async fn run_server(
             .service(activitypub::tag_view)
             .service(activitypub::conversation_view)
             .service(activitypub::activity_view)
-            .service(activitypub::gateway_scope())
             .service(atom_scope())
             .service(nodeinfo::get_nodeinfo_jrd)
             .service(nodeinfo::get_nodeinfo_2_0)
@@ -143,6 +142,9 @@ pub async fn run_server(
                 web::resource("/.well-known/{path}")
                     .to(HttpResponse::NotFound)
             );
+        if config.federation.fep_ef61_gateway_enabled {
+            app = app.service(activitypub::gateway_scope());
+        };
         #[allow(irrefutable_let_patterns)]
         if let MediaStorage::Filesystem(ref backend) = media_storage {
             app = app.service(actix_files::Files::new(