+++ mitra_config/src/federation.rs
@@ -12,6 +12,7 @@ const fn default_deliverer_log_response_length() -> usize { 75 }
 const fn default_fep_e232_enabled() -> bool { true }
 const fn default_fep_1b12_full_enabled() -> bool { true }
 const fn default_fep_c0e0_emoji_react_enabled() -> bool { false }
+const fn default_fep_ef61_gateway_enabled() -> bool { false }
 
 #[derive(Clone, Deserialize)]
 pub struct FederationConfig {
@@ -54,6 +55,9 @@ pub struct FederationConfig {
 
     #[serde(default = "default_fep_c0e0_emoji_react_enabled")]
     pub fep_c0e0_emoji_react_enabled: bool,
+
+    #[serde(default = "default_fep_ef61_gateway_enabled")]
+    pub fep_ef61_gateway_enabled: bool,
 }
 
 impl Default for FederationConfig {
@@ -74,6 +78,7 @@ impl Default for FederationConfig {
             fep_e232_enabled: default_fep_e232_enabled(),
             fep_1b12_full_enabled: default_fep_1b12_full_enabled(),
             fep_c0e0_emoji_react_enabled: default_fep_c0e0_emoji_react_enabled(),
+            fep_ef61_gateway_enabled: default_fep_ef61_gateway_enabled(),
         }
     }
 }